fix a very subtle bug related to float precision
In the special case when `rest_timeout` is 0, the bug would cause the
"stopped" state to alternate between true and false rapidly, preventing
any movement. This only happens after some time has elapsed, though,
because the error manifests only when `now_us` is sufficiently large
(> 12e6). In that case, even though `now_us` is equal to `cvel->now_us`,
their difference is not exactly 0 which causes the If to match.