implement velocity passthrough mode
previously, we could "disable" the position controller by calling
control_position_stop(). When disabled like this, the position
controller ramps down to velocity 0 and stays there.

we now extend this functionality to be able to disable the position
controller and keep a velocity *other than 0*, using the
control_position_ptru_mode() function. Calling it will cause the
controller's output to slowly ramp to the given velocity and stay there.