fix behaviour of position controller when coast speed is slow
When using the position controller with vel_coast < vel_min, the vel
output gets clamped to either 0 or -0 depending on direction. Due to
incorrect handling of the -0 value in some math util functions, this
resulted in the position controller never stopping.

The collateral damage of this is one (1) kitchen sink damaged by a
robot who went on an unstoppable murder spree at a very slow speed.